bugs fixed

+++ posts/views.py
@@ -10,7 +10,7 @@ def index(request):
     return render(request, 'posts/index.html')
 
 
-@login_required(redirect_field_name='/posts/')
+@login_required(login_url='/posts/')
 def post_create(request):
     post_form = PostForm()
 
@@ -26,11 +26,18 @@ def post_create(request):
 
 
 def post_list(request):
-    if request.GET.get('ingredient', None) is None:
-        posts = Post.objects.annotate(avg_rating=Avg('rate__score'))
-    else:
-        ingredient = Ingredients.objects.get(pk=request.GET.get('ingredient'))
-        posts = ingredient.posts.annotate(avg_rating=Avg('rate__score'))
+    posts = Post.objects.annotate(avg_rating=Avg('rate__score'))
+
+    pk = request.GET.get('ingredient', None)
+    if pk is not None:
+        try:
+            if Ingredients.objects.filter(pk=pk).exists():
+                ingredient = Ingredients.objects.get(pk=pk)
+                posts = posts.filter(id__in=ingredient.posts.values('id'))
+            else:
+                return HttpResponseRedirect(reverse('posts:post_list'))
+        except ValueError:
+            return HttpResponseRedirect(reverse('posts:post_list'))
 
     ingredients = Ingredients.objects.annotate(num_posts=Count('posts')).order_by('-num_posts')[:5]
 
@@ -43,7 +50,7 @@ def post_detail(request, pk):
 
     if request.user.is_authenticated:
         like = Like.objects.filter(post=post, user=request.user)
-        if Rate.objects.filter(post=post, user=request.user):
+        if Rate.objects.filter(post=post, user=request.user).exists():
             rate = Rate.objects.get(post=post, user=request.user)
             rate_form = RateForm(instance=rate)
     else:
@@ -52,7 +59,7 @@ def post_detail(request, pk):
     return render(request, 'posts/post_detail.html', context={'post': post, 'like': like, 'rate_form': rate_form})
 
 
-@login_required(redirect_field_name='/posts/')
+@login_required(login_url='/posts/')
 def post_edit(request, pk):
     if request.method == 'POST':
         post = Post.objects.get(pk=pk)
@@ -64,13 +71,26 @@ def post_edit(request, pk):
     return HttpResponseRedirect(reverse('posts:detail', kwargs={'pk': pk}))
 
 
-@login_required(redirect_field_name='/posts/')
-def update(request, pk):
+@login_required(login_url='/posts/')
+def post_delete(request, pk):
     if request.method == 'POST':
         post = Post.objects.get(pk=pk)
 
         if post.owner == request.user:
-            form = PostForm(data=request.POST, instance=post)
+            post.delete()
+        else:
+            raise PermissionDenied
+
+    return HttpResponseRedirect(reverse('posts:post_list'))
+
+
+@login_required(login_url='/posts/')
+def handle_update(request, pk):
+    if request.method == 'POST':
+        post = Post.objects.get(pk=pk)
+
+        if post.owner == request.user:
+            form = PostForm(data=request.POST, files=request.FILES, instance=post)
 
             if form.is_valid():
                 form.save(commit=True)
@@ -83,63 +103,47 @@ def update(request, pk):
     return HttpResponseRedirect(reverse('posts:detail', kwargs={'pk': pk}))
 
 
-@login_required(redirect_field_name='/posts/')
-def delete(request, pk):
-    if request.method == 'POST':
-        post = Post.objects.get(pk=pk)
-
-        if post.owner == request.user:
-            post.delete()
-        else:
-            raise PermissionDenied
-
-    return HttpResponseRedirect(reverse('posts:post_list'))
-
-
 def handle_like(request, pk):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.method == 'POST':
         post = Post.objects.get(pk=pk)
 
-        if request.method == 'POST':
-            if not Like.objects.filter(post=post, user=request.user):
-                # user likes the post if he/she hasn't liked it before
-                like = Like(post=post, user=request.user)
-                like.save()
+        if not Like.objects.filter(post=post, user=request.user).exists():
+            # user likes the post if he/she hasn't liked it before
+            like = Like(post=post, user=request.user)
+            like.save()
 
     return HttpResponseRedirect(reverse('posts:detail', kwargs={'pk': pk}))
 
 
 def handle_unlike(request, pk):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.method == 'POST':
         post = Post.objects.get(pk=pk)
 
-        if request.method == 'POST':
-            if Like.objects.filter(post=post, user=request.user):
-                # user unlikes the post if he/she has liked it before
-                like = Like.objects.get(post=post, user=request.user)
-                like.delete()
+        if Like.objects.filter(post=post, user=request.user).exists():
+            # user unlikes the post if he/she has liked it before
+            like = Like.objects.get(post=post, user=request.user)
+            like.delete()
 
     return HttpResponseRedirect(reverse('posts:detail', kwargs={'pk': pk}))
 
 
 def handle_rate(request, pk):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.method == 'POST':
         post = Post.objects.get(pk=pk)
 
-        if request.method == 'POST':
-            if Rate.objects.filter(post=post, user=request.user):
-                rate = Rate.objects.get(post=post, user=request.user)
-                form = RateForm(data=request.POST, instance=rate)
+        if Rate.objects.filter(post=post, user=request.user).exists():
+            rate = Rate.objects.get(post=post, user=request.user)
+            form = RateForm(data=request.POST, instance=rate)
 
-                if form.is_valid():
-                    form.save(commit=True)
-            else:
-                form = RateForm(request.POST, request.FILES)
-                if form.is_valid():
-                    rate = form.save(commit=False)
-                    rate.user = request.user
-                    rate.post = post
-                    rate.save()
+            if form.is_valid():
+                form.save(commit=True)
+        else:
+            form = RateForm(request.POST)
+            if form.is_valid():
+                rate = form.save(commit=False)
+                rate.user = request.user
+                rate.post = post
+                rate.save()
 
     return HttpResponseRedirect(reverse('posts:detail', kwargs={'pk': pk}))
